requirements.txt
kiwipiepy>=0.17.0
scikit-learn>=1.3.0
keybert>=0.8.0
sentence-transformers>=2.2.0
pytrends>=4.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
google-generativeai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.10.0
